+++ src/Spago/Build.hs
@@ -3,7 +3,6 @@ module Spago.Build
   , test
   , run
   , repl
-  , bundleApp
   , bundleModule
   , docs
   , search
@@ -42,11 +41,13 @@ import qualified Spago.Watch          as Watch
 prepareBundleDefaults
   :: Maybe ModuleName
   -> Maybe TargetPath
-  -> (ModuleName, TargetPath)
-prepareBundleDefaults maybeModuleName maybeTargetPath = (moduleName, targetPath)
+  -> Maybe Platform
+  -> (ModuleName, TargetPath, Platform)
+prepareBundleDefaults maybeModuleName maybeTargetPath maybePlatform = (moduleName, targetPath, platform)
   where
     moduleName = fromMaybe (ModuleName "Main") maybeModuleName
     targetPath = fromMaybe (TargetPath "index.js") maybeTargetPath
+    platform = fromMaybe Browser maybePlatform
 
 --   eventually running some other action after the build
 build :: HasBuildEnv env => Maybe (RIO Env ()) -> RIO env ()
@@ -374,46 +375,66 @@ runBackend maybeBackend RunDirectories{ sourceDir, executeDir } moduleName maybe
         ExitSuccess   -> maybe (pure ()) (logInfo . display) maybeSuccessMessage
         ExitFailure n -> die [ display failureMessage <> "Backend " <> displayShow backend <> " exited with error:" <> repr n ]
 
--- | Bundle the project to a js file
-bundleApp
-  :: HasEnv env
-  => WithMain
-  -> Maybe ModuleName
-  -> Maybe TargetPath
-  -> NoBuild
-  -> BuildOptions
-  -> UsePsa
-  -> RIO env ()
-bundleApp withMain maybeModuleName maybeTargetPath noBuild buildOpts usePsa =
-  let (moduleName, targetPath) = prepareBundleDefaults maybeModuleName maybeTargetPath
-      bundleAction = Purs.bundle withMain (withSourceMap buildOpts) moduleName targetPath
-  in case noBuild of
-    DoBuild -> Run.withBuildEnv usePsa buildOpts $ build (Just bundleAction)
-    NoBuild -> Run.getEnv >>= (flip runRIO) bundleAction
 
--- | Bundle into a CommonJS module
+bundleWithEsbuild :: HasLogFunc env => WithMain -> ModuleName -> TargetPath -> Platform -> Minify -> RIO env ()
+bundleWithEsbuild withMain (ModuleName moduleName) (TargetPath targetPath) platform minify = do
+  esbuild <- getESBuild
+  let
+    platformOpt = case platform of
+      Browser -> "browser"
+      Node -> "node"
+    minifyOpt = case minify of
+      NoMinify -> ""
+      Minify -> " --minify"
+    cmd = case withMain of
+      WithMain ->
+        "echo \"import { main } from './output/" <> moduleName <> "/index.js'\nmain()\" | "
+        <> esbuild <> " --platform=" <> platformOpt <> minifyOpt <> " --bundle "
+        <> " --outfile=" <> targetPath
+      WithoutMain ->
+        esbuild <> " --platform=" <> platformOpt <> minifyOpt <> " --bundle "
+        <> "output/" <> moduleName <> "/index.js"
+        <> " --outfile=" <> targetPath
+  runWithOutput cmd
+    ("Bundle succeeded and output file to " <> targetPath)
+    "Bundle failed."
+  where
+  getESBuild = do
+    maybeESBuild <- findExecutable "esbuild"
+    case maybeESBuild of
+      Nothing -> die [ "Failed to find esbuild. See https://esbuild.github.io/getting-started/#install-esbuild for ways to install esbuild." ]
+      Just esBuild -> pure esBuild
+
+-- | Bundle the project to a CommonJs module/app or an ES module
 bundleModule
-  :: HasEnv env
-  => Maybe ModuleName
-  -> Maybe TargetPath
-  -> NoBuild
+  :: (HasEnv env, HasPurs env)
+  => WithMain
+  -> BundleOptions
   -> BuildOptions
   -> UsePsa
   -> RIO env ()
-bundleModule maybeModuleName maybeTargetPath noBuild buildOpts usePsa = do
-  logDebug "Running `bundleModule`"
-  let (moduleName, targetPath) = prepareBundleDefaults maybeModuleName maybeTargetPath
-      jsExport = Text.unpack $ "\nmodule.exports = PS[\""<> unModuleName moduleName <> "\"];"
-      bundleAction = do
-        logInfo "Bundling first..."
-        Purs.bundle WithoutMain (withSourceMap buildOpts) moduleName targetPath
-        -- Here we append the CommonJS export line at the end of the bundle
-        try (with
-              (appendonly $ pathFromText $ unTargetPath targetPath)
-              (\fileHandle -> Utf8.withHandle fileHandle (Sys.hPutStrLn fileHandle jsExport)))
-          >>= \case
-            Right _ -> logInfo $ display $ "Make module succeeded and output file to " <> unTargetPath targetPath
-            Left (n :: SomeException) -> die [ "Make module failed: " <> repr n ]
+bundleModule withMain BundleOptions { maybeModuleName, maybeTargetPath, maybePlatform, minify, noBuild }  buildOpts usePsa = do
+  isES <- Purs.hasMinPursVersion "0.15.0"
+  let
+    (moduleName, targetPath, platform) = prepareBundleDefaults maybeModuleName maybeTargetPath maybePlatform
+    bundleAction =
+      if isES then
+        bundleWithEsbuild withMain moduleName targetPath platform minify
+      else case withMain of
+        WithMain -> Purs.bundle WithMain (withSourceMap buildOpts) moduleName targetPath
+        WithoutMain ->
+          let
+            jsExport = Text.unpack $ "\nmodule.exports = PS[\""<> unModuleName moduleName <> "\"];"
+          in do
+              logInfo "Bundling first..."
+              Purs.bundle WithoutMain (withSourceMap buildOpts) moduleName targetPath
+              -- Here we append the CommonJS export line at the end of the bundle
+              try (with
+                    (appendonly $ pathFromText $ unTargetPath targetPath)
+                    (\fileHandle -> Utf8.withHandle fileHandle (Sys.hPutStrLn fileHandle jsExport)))
+                >>= \case
+                  Right _ -> logInfo $ display $ "Make module succeeded and output file to " <> unTargetPath targetPath
+                  Left (n :: SomeException) -> die [ "Make module failed: " <> repr n ]
   case noBuild of
     DoBuild -> Run.withBuildEnv usePsa buildOpts $ build (Just bundleAction)
     NoBuild -> Run.getEnv >>= (flip runRIO) bundleAction